Ender - Nice
Made a small tweak to my conquest function that only sends units to bases for takeover if their arrival will be on a round ending in a 9. 
Seems small, but if you have to send 2 units and they would arrive on a tick ending in 1, they will stand around for 8 rounds waiting to takeover...

Thats 8 rounds they could technically have been standing around burning wood.